+++ src/com/vv0rkman/entity/Entity.java
@@ -7,23 +7,23 @@ public class Entity implements Serializable {
     protected static final int PARENT_ID = 2;
     protected static final int NAME = 1;
 
-    int id;
+    Long id;
     int parent_id;
     String name;
 
     Entity(){}
 
-    Entity(int id, int parent_id, String name) {
+    Entity(Long id, int parent_id, String name) {
         this.id = id;
         this.parent_id = parent_id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
